Consolidate exchangeQuoteProcessing error handling into scene component
